+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/RotateParticleSpringboneRuntimeProvider.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+using UniVRM10;
+
+namespace RotateParticle
+{
+    public class RotateParticleSpringboneRuntimeProvider : MonoBehaviour, IVrm10SpringBoneRuntimeProvider
+    {
+        public IVrm10SpringBoneRuntime CreateSpringBoneRuntime()
+        {
+            return new RotateParticleSpringboneRuntime();
+        }
+    }
+}
