+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/RotateParticleSpringboneRuntimeProvider.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f2c23a62e016b3244b286404bc339540
+guid: de8d095eee60f134ebd9ccaa9efcc01e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
